+++ test/cbmc/proofs/MQTT_SerializeSubscribe/MQTT_SerializeSubscribe_harness.c
@@ -33,9 +33,9 @@ void harness()
 {
     MQTTSubscribeInfo_t * pSubscriptionList;
     size_t subscriptionCount;
-    size_t remainingLength;
+    uint32_t remainingLength;
     uint16_t packetId;
-    size_t packetSize;
+    uint32_t packetSize;
     MQTTFixedBuffer_t * pFixedBuffer;
     MQTTStatus_t status = MQTTSuccess;
     MQTTPropBuilder_t * pSubscribeProperties;
